Darken Polar Night palette to increase theme contrast
First off, the Nord theme is awesome. I've been using it for years and
really appreciate how well thought out the entire project is. So why
this fork? Well, one thing that sometimes bothers my eyes is the low
contrast. So this is my attempt to darken Polar Night to increase the
text contrast while maintaining the integrity of the theme as a whole.

Polar Night has four shades of gray. So to darken the palette, I
introduced a new shade for the darkest color, `#242933`, which was
pulled from the background of the Nord (dark) website. Then each of the
four colors shifted over by one, dropping the brightest gray.

```sh
replace '#2E3440' '#242933'
replace '#3B4252' '#2E3440'
replace '#434C5E' '#3B4252'
replace '#4C566A' '#434C5E'
```